--- a/sky/backends/backend_utils.py
+++ b/sky/backends/backend_utils.py
@@ -839,6 +839,13 @@ def write_cluster_config(
 
     logger.debug(f'Using ssh_proxy_command: {ssh_proxy_command!r}')
 
+    disk_configure = {
+        'disk_type': to_provision.get_cloud_disk_type_str(),
+    }
+    # AWS need to specify iops and throughput manually.
+    if isinstance(cloud, clouds.AWS):
+        disk_configure['disk_iops'] = to_provision.get_disk_iops()
+
     # Use a tmp file path to avoid incomplete YAML file being re-used in the
     # future.
     tmp_yaml_path = yaml_path + '.tmp'
@@ -898,6 +905,8 @@ def write_cluster_config(
                 'worker_ips': None if ip_list is None else ip_list[1:],
                 # Authentication (optional).
                 **auth_config,
+                # Disk configuration.
+                **disk_configure,
             }),
         output_path=tmp_yaml_path)
     config_dict['cluster_name'] = cluster_name