--- a/sky/backends/backend_utils.py
+++ b/sky/backends/backend_utils.py
@@ -850,7 +850,6 @@ def write_cluster_config(
                 'cluster_name': cluster_name,
                 'num_nodes': num_nodes,
                 'disk_size': to_provision.disk_size,
-                'disk_type': to_provision.get_cloud_disk_type(),
                 # If the current code is run by controller, propagate the real
                 # calling user which should've been passed in as the
                 # SKYPILOT_USER env var (see spot-controller.yaml.j2).