--- a/sky/cli.py
+++ b/sky/cli.py
@@ -3305,7 +3305,7 @@ def spot():
               default=None,
               type=str,
               required=False,
-              help=('OS disk type. Could be one of "low", "medium", "high".'))
+              help=('OS disk type. Could be one of "low", "medium", "high". Default: medium'))
 @click.option(
     '--detach-run',
     '-d',