… docs/source/reference/yaml-spec.rst
… sky/clouds/azure.py
… sky/clouds/cloud.py
… sky/clouds/gcp.py
… sky/clouds/lambda_cloud.py
… sky/clouds/local.py
… sky/clouds/service_catalog/__init__.py
… sky/clouds/service_catalog/aws_catalog.py
… sky/clouds/service_catalog/azure_catalog.py
… sky/clouds/service_catalog/gcp_catalog.py
… sky/clouds/service_catalog/lambda_catalog.py
… sky/resources.py
… sky/skylet/providers/azure/azure-vm-template.json
… sky/templates/aws-ray.yml.j2
… sky/templates/azure-ray.yml.j2
… sky/templates/gcp-ray.yml.j2
… sky/templates/spot-controller.yaml.j2
… sky/utils/schemas.py
… tests/test_optimizer_dryruns.py
… tests/test_smoke.py