<!-- Describe the changes in this PR -->

This PR introduces multi-level performance disks in all clouds. By specifying `disk_tier` in `sky.Resources` to one of `high`, `medium`, and `low`, a user could use a disk with custom performance. All clouds' current decisions and prices (256 GB for example) are shown below.

`disk_tier=low`
|Cloud|Disk Type|Benchmarked Read Throughput|Benchmarked Read IOPS|Benchmarked Write Throughput|Benchmarked Write IOPS|Price (GiB/mo)|Price (IOPS/mo)|Total Price Each Month|
|-|-|-|-|-|-|-|-|-|
|GCP|pd-standard|13.87 MB/s|211.63|37.17 MB/s|567.13|0.048|-|12.288|
|Azure|Standard_LRS|861.92 MB/s|13151.92|36.67 MB/s|559.54|-|-|11.328|
|AWS|standard|20.33 MB/s|310.14|282.19 MB/s|4305.91|0.066|-|16.896|

`disk_tier=medium`
|Cloud|Disk Type|Benchmarked Read Throughput|Benchmarked Read IOPS|Benchmarked Write Throughput|Benchmarked Write IOPS|Price (GiB/mo)|Price (IOPS/mo)|Total Price Each Month|
|-|-|-|-|-|-|-|-|-|
|GCP|pd-balanced|222.86 MB/s|3400.58|222.25 MB/s|3391.25|0.12|-|30.72|
|Azure|Premium_LRS|211.80 MB/s|3231.88|176.12 MB/s|2687.44|-|-|36.29|
|AWS|gp3 with IOPS=3500|241.43 MB/s|3683.87|199.15 MB/s|3038.86|0.09|0.005|25.54|

`disk_tier=high`
|Cloud|Disk Type|Benchmarked Read Throughput|Benchmarked Read IOPS|Benchmarked Write Throughput|Benchmarked Write IOPS|Price (GiB/mo)|Price (IOPS/mo)|Total Price Each Month|
|-|-|-|-|-|-|-|-|-|
|GCP|pd-ssd|342.91 MB/s|5232.42|268.17 MB/s|4091.91|0.203|-|51.968|
|Azure|-|-|-|-|-|-|-|-|
|AWS|gp3 with IOPS=7000|323.54 MB/s|4936.80|254.82 MB/s|3888.22|0.09|0.005|43.04|

Azure `disk_tier=high` is disabled now since a high-performance disk in Azure cannot be launched as an OS disk.

Benchmark command:

```shell
fio --name=sync_rand_64k_{read, write}s  --rw=rand{read, write}  --direct=1 --ioengine=sync --bs=64k --numjobs=4 --iodepth=128 --size=1G --group_reporting --directory=/tmp/
```

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [ ] Any manual or new tests for this PR (please specify below)
- [ ] All smoke tests: `pytest tests/test_smoke.py` 
- [ ] Relevant individual smoke tests: `pytest tests/test_smoke.py::test_fill_in_the_name` 
- [ ] Backward compatibility tests: `bash tests/backward_comaptibility_tests.sh`