@concretevitamin 2023-04-09
Skimming the description, this looks great! One quick thing I noticed is `disk_type=low` maps to very different performance on the three clouds. Is this desirable or by design? 

Asking because previous users have found that the default GCP disk is too slow. If we default to `medium`, this may be fine.