@Michaelvll 2023-04-09
> Skimming the description, this looks great! One quick thing I noticed is `disk_type=low` maps to very different performance on the three clouds. Is this desirable or by design?

Please correct me if I am wrong @cblmemo. That is the best compromisation we can get to match the performance among all three clouds, as the worst disks on those clouds have very different performances.
It should be fine, as we just guarantee the lower bound of the disk performance.

> Asking because previous users have found that the default GCP disk is too slow. If we default to `medium`, this may be fine.

Yep, good idea. Defaulting to medium is a good choice.