@cblmemo 2023-04-10
> Thanks for the quick fix @cblmemo! The latest code looks good to me. I am testing it for now.
> 
> Several questions:
> 
> 1. should we update `sky/utils/schemas.py` to ensure the `resources` field can take the `disk_tier` argument?
> 2. If we have the current PR merged, will [[Optimizer] Add GCP disk price to the optimizer #1708](https://github.com/skypilot-org/skypilot/pull/1708), [[Optimizer] Add Azure disk price to the optimizer #1744](https://github.com/skypilot-org/skypilot/pull/1744) need to be updated?

For the second question, yes, we need to add multi-tier prices into the disk price calculation. The previous implementation just fetches the default tier disk type's price and uses it to calculate the total price.