@Michaelvll 2023-04-16
Thanks for the awesome work and the quick fixes @cblmemo! The PR looks good to me and I also tested it with the smoke tests on GCP, manually tried several combinations of the options with the disk tier, and launched a VM on azure. They all look good! Merging the PR now.